requirements.txt
firebase-admin==6.4.0
python-dotenv==1.0.1
flask==3.0.2 